iot_device/requirements.txt
AWSIoTPythonSDK==1.5.4
RPi.GPIO
numpy
scikit-learn
joblib
pandas
matplotlib
seaborn
